docker_app/requirements.txt
opencv-python==4.9.0.80
rembg==2.0.53
boto3==1.34.29
matplotlib==3.8.2
python-dotenv==1.0.1
streamlit==1.30.0